+++ quam_builder/builder/qop_connectivity/__init__.py
@@ -1,2 +1,11 @@
 from .build_quam_wiring import build_quam_wiring
-from .modify_ports import add_port, remove_port
+from .modify_ports import (
+    add_analog_port,
+    add_digital_port,
+    add_mw_port,
+    add_port,
+    remove_analog_port,
+    remove_digital_port,
+    remove_mw_port,
+    remove_port,
+)